+++ src/modelo/Controle.java
@@ -1,13 +1,8 @@
 package modelo;
 
-public class Controle
+public class Controle extends absPropriedades
 {
-    private String num1;
-    private String num2;
-    private String mensagem;
-    private String operacao;
-    private String resultado;
-
+    
     public Controle(String num1, String num2, String operacao)
     {
         this.num1 = num1;
@@ -33,18 +28,5 @@ public class Controle
         {
             this.mensagem = validacao.getMensagem();
         }
-        
-        
-        
-    }
-
-    public String getMensagem()
-    {
-        return mensagem;
-    }
-
-    public String getResultado()
-    {
-        return resultado;
     }
 }
