+++ src/modelo/absPropriedades.java
@@ -1,0 +1,35 @@
+package modelo;
+
+
+
+public class absPropriedades
+{
+    public Double numero1;
+    public Double numero2;
+    public String operacao;
+    public String resultado;
+    public String num1;
+    public String num2;
+    public String mensagem;
+    
+    
+    public String getResultado()
+    {
+        return resultado;
+    }
+    
+    public String getMensagem()
+    {
+        return mensagem;
+    }
+
+    public Double getNumero1()
+    {
+        return numero1;
+    }
+
+    public Double getNumero2()
+    {
+        return numero2;
+    }
+}
